backend/requirements.txt
flask
flask-cors
pinecone-client
sentence-transformers
openai
requests